implement index files for command handlers
